+++ src/main/java/me/shepherd23333/projectex/client/ProjectEXClient.java
@@ -5,6 +5,8 @@ import me.shepherd23333.projectex.integration.PersonalEMC;
 import net.minecraft.client.Minecraft;
 import net.minecraft.entity.player.EntityPlayer;
 
+import java.math.BigInteger;
+
 /**
  * @author LatvianModder
  */
@@ -15,7 +17,7 @@ public class ProjectEXClient extends ProjectEXCommon {
     }
 
     @Override
-    public void updateEMC(long emc) {
+    public void updateEMC(BigInteger emc) {
         PersonalEMC.get(Minecraft.getMinecraft().player).setEmc(emc);
     }
 }
